arm: dts: qcom: Add device tree for Microsoft Lumia 435
Microsoft Lumia 435 (vela) is a smartphone using the MSM8210 SoC.

This commit adds a DT for this phone with initial support for:
  - SDHCI
  - Screen via EFIFB
  - I2C
  - Touchscreen

All other MSM8210 Lumia phones are very similar, with some differences in display, touchscreen
and sensors. The common parts are shared in
qcom-msm8210-msft-lumia-common.dtsi to reduce duplication.

Change-Id: I3e3c83e0e58b7312992ffab462504aa981b19533